feature: 新增构件分发功能到服务 #902
