+++ src/backend/ci/core/artifactory/biz-artifactory-push/src/main/kotlin/com/tencent/devops/artifactory/service/FileService.kt
@@ -1,0 +1,13 @@
+package com.tencent.devops.artifactory.service
+
+import java.io.File
+
+interface FileService {
+    fun downloadFileTolocal(
+        projectId: String,
+        pipelineId: String,
+        buildId: String,
+        fileName: String,
+        isCustom: Boolean
+    ): List<File>
+}
